requirements.txt
eclipse-zenoh
edgefirst-schemas
numpy
rerun-sdk
zstd
edgefirst-client
pillow